Move clang-format.el along with tools-extra (bz#1812487)
With the current setup, if you only install clang-tools-extra, the
package clang-format is not found by Emacs.

A similar change was made for the Python scripts, although I did not
run in the same situation myself.

v2: Moved additional files at the suggestion of @sergesanspaille
%{_mandir}/man1/diagtool.1.gz
%{_bindir}/c-index-test
Also restored trailing whitespaces in the .spec file