Describe weak dependencies
clang has a couple of indirect weak dependencies; installing clang
without those will cause it to fail in certain use cases, but its
description doesn't mention them which makes resolving this harder
than it could be.

This patch adds a brief description of the weak dependencies to the
main clang package description.

Closes: #1879410
Signed-off-by: Stephen Kitt <skitt@redhat.com>