Backport a patch from LLVM trunk
Since LLVM 17.0.6 is the last release of the 17.x branch. the fix backported
by this change helps to address a critical build failure with libstdc++
shipped with GCC-13, so we need to include it in the clang packages.

the include patch was updated to resolve the conflicts when being
applied to the llvmorg-17.0.6 tag.

the fix has been tested after applying it on top of upstream's
llvmorg-17.0.6 tag.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>